back/pyproject.toml
[tool.poetry]
name = "back"
version = "0.1.0"
description = ""
authors = ["Igor4er <sherednihenko.ihor@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.100.1"
peewee = "^3.16.2"
python-dotenv = "^1.0.0"
pydantic-settings = "^2.0.2"
uvicorn = "^0.23.2"
pyjwt = "^2.8.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
